Create InplaceVector
C++ has two main contiguous collections which are std::vector and std::array. std::vector provides a heap allocated growable array with a fast move operation while std::array is an in place fixed size wrapper over C-style arrays that handles object lifetimes correctly. While these two cover most use cases pretty adequately, there is the issue of small non-fixed size arrays. There are some cases where we need resizable arrays of a fairly small size. It is possible to emulate this if you have a sentinel element with std::array but I find this to be too error-prone. Falling back to std::vector comes with performance downsides. Although modern allocators are quite fast, having a lot of small allocations tend to waste a lot of heap space that can degrade performance over time. This becomes especially clear when you attempt to multithread existing code where previously alright allocator performance is suddenly degraded due to an increase in contention. People have tried to work around this by using alternative allocator like jemalloc however this is more of a crutch than a solution. The problem becomes more pronounced when we want to have a "vector of vectors". This is almost the worst case scenario for indirection with the CPU having little idea where the memory accesses will be but is also the worst case scenario for memory fragmentation which means worse performance due to cache misses and significantly increased memory usage. There are already use cases within GEL where creating a vector of 300,000+ vectors is not unusual, so there is a clear use case for this class.

Note that this is not a "SmallVector" which is often used to indicate a growable array that can be conditionally stored in place below a minimum size but can also be moved to the heap beyond that size.

If the library is ever updated to C++26, this class can likely be replaced using std::inplace_vector, which provides std::vector semantics with a fixed maximum capacity.