Add Threadpool implementation
These are threadpool implementations for parallel adapters that are used within the RSR algorithm. I decided to make this separate so I can easily swap out implementations which can be useful during profiling as multithreading can sometimes make results more ambiguous. In addition, we cannot use the built-in parallel algorithms. Not only is C++20 missing a good number of algorithms that we are interested in using (or have variants that are undesirable to use in some way), but one of the compilers that we wish to support (Apple Clang) does not support execution policies at all. Even when execution policies are supported, the C++ standard places no requirements on implementors to actually execute relevant algorithms in parallel when parallel execution is requested.

The ThreadPool class seems to still contain a subtle race condition where the pool can deadlock if the given closure finishes really quickly. I initially wanted to use this design because it amortizes thread creation costs but I might end up just removing it as our algorithms take long enough that thread creation is typically completely negligible.