Add Parallel Adapters
Adds concepts and functions for parallel algorithms.

These add parallel algorithms comparable to ones available in the algorithm header. Note that execution policy is still not supported on Apple Clang right now (2025-07-22) which just so happens to be the compiler Andreas uses.

This library is much stricter than the algorithms one in the sense that we require the input and output ranges to be contiguous and have constant time random access. While there are more generic ways of writing parallel algorithms, cache locality can lose us more performance (>10x) than what we gain thanks to multithreading.

Also note that the names follow functional programming concepts rather than the C++ std names (map, filter instead of transform, erase_if) since the choice of some of the std names were picked to either avoid collisions with existing portions of the library or follow some specific C++ idiom which we don't necessarily care about.

It is also worth noting that there is quite a bit of duplication here to express some more complex adapters that would normally be composed in more functional code. This is because we need to block to force computations to complete. Allowing composition without blocking would require a Rust style library where computation is indirectly described in terms of traits and an explicit collect is added to the end to finish the computation. It is possible to write this in C++ but it is overly complex and most of the computation that we are doing can be described in terms of a simple map-fold.

Note that I might also turn IExecutor into a template parameter to allow the compiler to inline one more level although I suspect the impact of thi son performance is negligible.